get_loss_and_metric moved under "parameterParsing"
